pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.2",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "attrs==25.4.0",
    "certifi==2025.10.5",
    "charset-normalizer==3.4.4",
    "colorama==0.4.6 ; sys_platform == 'win32'",
    "dataclasses-json==0.6.7",
    "faiss-cpu==1.12.0",
    "filelock==3.20.0",
    "frozenlist==1.8.0",
    "fsspec==2025.10.0",
    "greenlet==3.2.4 ; platform_machine == 'AMD64' or platform_machine == 'WIN32' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'ppc64le' or platform_machine == 'win32' or platform_machine == 'x86_64'",
    "h11==0.16.0",
    "hf-xet==1.2.0 ; platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'arm64' or platform_machine == 'x86_64'",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "huggingface-hub==0.36.0",
    "idna==3.11",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==1.0.5",
    "langchain-classic==1.0.0",
    "langchain-community==0.4.1",
    "langchain-core==1.0.4",
    "langchain-google-genai>=3.0.1",
    "langchain-huggingface==1.0.1",
    "langchain-text-splitters==1.0.0",
    "langgraph==1.0.2",
    "langgraph-checkpoint==3.0.1",
    "langgraph-prebuilt==1.0.2",
    "langgraph-sdk==0.2.9",
    "langsmith==0.4.42",
    "marshmallow==3.26.1",
    "multidict==6.7.0",
    "mypy-extensions==1.1.0",
    "numpy==2.3.4",
    "orjson==3.11.4",
    "ormsgpack==1.12.0",
    "packaging==25.0",
    "propcache==0.4.1",
    "pydantic==2.12.4",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.12.0",
    "python-dotenv==1.2.1",
    "pyyaml==6.0.3",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "sentence-transformers>=5.1.2",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.44",
    "streamlit>=1.51.0",
    "tenacity==9.1.2",
    "tokenizers==0.22.1",
    "tqdm==4.67.1",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "urllib3==2.5.0",
    "xxhash==3.6.0",
    "yarl==1.22.0",
    "zstandard==0.25.0",
]
